Move dropdown alignment and context key to seperated file
