+++ attractions/dropdown/dropdown-context-key.js
@@ -1,0 +1,2 @@
+export const CONTEXT_GET_DROPDOWN_SHELL_BOUNDARY = 'getDropdownShellBoundary';
+export const CONTEXT_IS_DROPDOWN_OPEN = 'isDropdownOpen';
