+++ attractions/dropdown/index.js
@@ -1,0 +1,1 @@
+export { HorizontalAlignment, VerticalAlignment } from './dropdown-alignment';
